+++ imio/dms/mail/browser/static/dmsmail.js
@@ -55,6 +55,9 @@ dmsmail.initializeOverlays = function (form_id) {
 
 /* we have to copy this from imio.helpers to make it work in overlays */
 dmsmail.initialize_fancytree = function () {
+  if ($("form#tree-form").length == 0) {
+    return
+  }
   var submitButton = $("#tree-form input[type='submit'");
   var uidInput = $('#tree-form input[name="uid"]');
   var nodes = JSON.parse(document.getElementById('tree').dataset.nodes);
@@ -87,8 +90,8 @@ $(document).ready(function(){
     /* remove inline validation for dmsoutgoingmail
     $('.template-dmsoutgoingmail .z3cformInlineValidation, .template-dmsdocument-edit.portaltype-dmsoutgoingmail .z3cformInlineValidation').removeClass('z3cformInlineValidation'); */
 
-/*
-Added with first version of create form template on im r22564
+
+/* Added with first version of create form template on im r22564 */
     $('.overlay').prepOverlay({
         subtype: 'ajax',
         closeselector: '[name="form.buttons.cancel"]'
@@ -96,6 +99,6 @@ Added with first version of create form template on im r22564
 
     $(document).bind('loadInsideOverlay', function(e, el, responseText, errorText, api) {
         dmsmail.initialize_fancytree();
-    });*/
+    });
 
 });