+++ imio/dms/mail/tests/test_columns.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 import unittest
+from z3c.relationfield.relation import RelationValue
+from zope.component import getUtility
+from zope.intid.interfaces import IIntIds
 
 from plone import api
 from plone.app.testing import setRoles, TEST_USER_ID
@@ -16,6 +19,7 @@ class TestColumns(unittest.TestCase):
     def setUp(self):
         self.portal = self.layer['portal']
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.intids = getUtility(IIntIds)
         self.mail_table = self.portal['incoming-mail']['mail-searches'].unrestrictedTraverse('@@faceted-table-view')
         self.task_table = self.portal['incoming-mail']['mail-searches'].unrestrictedTraverse('@@faceted-table-view')
         self.imf = self.portal['incoming-mail']
@@ -32,12 +36,21 @@ class TestColumns(unittest.TestCase):
                          "class='pretty_link link-tooltip'><span class='pretty_link_icons'><img title='Held position' "
                          "src='http://nohost/plone/held_position_icon.png' /></span><span class='pretty_link_content'"
                          ">Monsieur Jean Courant (Electrabel - Agent)</span></a>")
+        # multiple senders
+        self.im5.sender.append(RelationValue(self.intids.getId(self.portal['contacts']['sergerobinet'])))
+        self.im5.reindexObject(idxs=['sender_index'])
+        brain = self.portal.portal_catalog(UID=self.im5.UID())[0]
+        rendered = column.renderCell(brain)
+        self.assertIn('<ul class="contact_list_col"><li>', rendered)
+        self.assertEqual(rendered.count('<a href'), 2)
         # no sender
         imail = createContentInContainer(self.imf, 'dmsincomingmail', id='my-id', title='My title',
                                          description='Description')
         brain = self.portal.portal_catalog(UID=imail.UID())[0]
         self.assertEqual(column.renderCell(brain), '-')
         # sender not found: we delete it
+        self.im5.sender = self.im5.sender[0:1]
+        self.im5.reindexObject(idxs=['sender_index'])
         api.content.delete(obj=self.portal['contacts']['jeancourant']['agent-electrabel'], check_linkintegrity=False)
         brain = self.portal.portal_catalog(UID=self.im5.UID())[0]
         self.assertEqual(column.renderCell(brain), '-')